refactor(iso): separate ISO builder concerns into dedicated modules
This commit modularizes the ISO builder implementation by extracting utility functions into separate modules. The changes include:
- Moving calculation and helper functions to builder_utils module
- Creating iso_writer module for ISO writing operations
- Reorganizing imports to reflect the new module structure
- Improving code organization and maintainability

This refactoring makes the codebase more modular and easier to understand while maintaining the same functionality.
