+++ src/iso/gpt/mod.rs
@@ -1,0 +1,3 @@
+pub mod header;
+pub mod partition_entry;
+pub mod main_gpt_functions;
