+++ src/iso/mod.rs
@@ -4,7 +4,9 @@ pub mod builder;
 pub mod constants;
 pub mod dir_record;
 pub mod fs_node;
-pub mod gpt;
+pub mod gpt; // Re-add this to make the gpt module accessible
 pub mod iso_image;
+pub mod builder_utils;
+pub mod iso_writer;
 pub mod mbr;
 pub mod volume_descriptor;
